restify-backend/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
Pillow==9.0.1
django-phonenumber-field==6.1.0
phonenumbers==8.12.44
django-cors-headers==3.11.0
python-dotenv==1.0.0